refactor(web): rename UploadForm.ts to ts and export type instead of export default
